fix: narrow browser memory saver policy updates
